+++ R/globals.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("int.col", "int.row","clusters","jc.extent"))
+utils::globalVariables(c("int.col","int.row","clusters","jc.extent","numericCluster","x","y","cuts"))