+++ R/joincount.R
@@ -1,16 +1,17 @@
-#' Join count analysis.
+#' This function performes multi-categorical join count analysis of
+#' the rasterized sample. A neighbors list is then created from the rasterized sample,
+#' in which adjacent and diagonal neighbors for each pixel are identified.
 #'
-#' @import Seurat
-#' @import spdep
+#' @importFrom spdep cell2nb nb2listw joincount.multi subset.nb
 #' @importFrom stats na.omit
 #' @export
 #'
-#' @param mosaicIntegration RasterLayer that output from Function rasterization.
+#' @param mosaicIntegration A raster object converted from a labeled spatial tissue map from Function rasterization.
 #'
-#' @return Join count analysis results.
+#' @return the observed join counts, the expected count under conditions of spatial randomness, the variance calculated under non-free sampling, and calculated Z-score.
 #'
 #' @examples
-#' fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
+#' fpath <- system.file("script", "humanBC.rda", package="stJoincount")
 #' load(fpath)
 #' mosaicIntegration <- rasterizeEachCluster(humanBC)
 #' joincount.result <- joincountAnalysis(mosaicIntegration)
@@ -22,7 +23,7 @@ joincountAnalysis <- function(mosaicIntegration){
   subList <- subset.nb(nbList,
                        !(1:length(nbList) %in% emptyPos))
   temp.subList <- lapply(subList, as.character)
-  noNeighborPos <- which(sapply(temp.subList, FUN =  function(X) "0" %in% X))
+  noNeighborPos <- which(vapply(temp.subList, FUN =  function(X) "0" %in% X, logical(1)))
 
   filteredList <- subset.nb(subList,
                             !(1:length(subList) %in% noNeighborPos))
@@ -42,11 +43,10 @@ joincountAnalysis <- function(mosaicIntegration){
                                           zero.policy = FALSE,
                                           adjust.n = FALSE)
   } else {
-    print("The length of mosaiced data and the length of weights list is not equal.")
+    stop("The length of mosaiced data and the length of weights list is not equal.")
   }
   joincount.result <- as.data.frame(joincount.filtered)
   joincount.result <- na.omit(joincount.result)
 
   return(joincount.result)
 }
-