+++ R/zscoreHeatmap.R
@@ -1,50 +1,51 @@
-#' Z-score matrix.
+#' This function provides a heatmap of z-scores resulting
+#' from the join count analysis for all possible label pairs.
 #'
-#' @import Seurat
-#' @import reshape2
+#' @param sampleInfo A dataset of a human breast cancer sample containing the
+#' pixel information and cluster labels for each barcode.
+#' @param joincount.result calculated result from join count analysis
 #' @export
 #'
-#' @param sample seruat object that have cluster labels attached
-#' @param joincount.result calculated result from join count analysis
-#'
-#' @return z-score matrix
+#' @return z-score matrix resulting from the join count analysis for all possible label pairs
 #'
 #' @examples
-#' fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
+#' fpath <- system.file("script", "humanBC.rda", package="stJoincount")
 #' load(fpath)
 #' mosaicIntegration <- rasterizeEachCluster(humanBC)
 #' joincount.result <- joincountAnalysis(mosaicIntegration)
 #' matrix <- zscoreMatrix(humanBC, joincount.result)
 
-zscoreMatrix <- function(sample, joincount.result){
-  clusterNumbers <- length(unique(sample$Cluster))
-  jcMatrix <- data.frame(matrix(NA, nrow = clusterNumbers, ncol = clusterNumbers))
-  colnames(jcMatrix) <- as.character(c(1:clusterNumbers))
-  rownames(jcMatrix) <- as.character(c(1:clusterNumbers))
+zscoreMatrix <- function(sampleInfo, joincount.result){
+    clusterNumbers <- length(unique(sampleInfo$Cluster))
+    jcMatrix <- data.frame(matrix(NA, nrow = clusterNumbers, ncol = clusterNumbers))
+    nameList <- customDict(sampleInfo)
+    clusterName <- nameList[seq_len(clusterNumbers)]
+    colnames(jcMatrix) <- clusterName
+    rownames(jcMatrix) <- clusterName
 
-  for (i in 1:clusterNumbers){
-    for (j in 1:clusterNumbers){
-      index1 <- paste(i, ':', j, sep = "")
-      jcMatrix[i,j] <- joincount.result[index1, 'z-value']
-      jcMatrix[j,i] <- joincount.result[index1, 'z-value']
+    for (i in seq_len(clusterNumbers)){
+      for (j in seq_len(clusterNumbers)){
+        index1 <- paste(i, ':', j, sep = "")
+        jcMatrix[i,j] <- joincount.result[index1, 'z-value']
+        jcMatrix[j,i] <- joincount.result[index1, 'z-value']
+      }
     }
-  }
-  jcMatrix <- round(jcMatrix, 2)
-  return(jcMatrix)
+    jcMatrix <- round(jcMatrix, 2)
+    return(jcMatrix)
 }
 
-
-#' Z-score heatmap.
+#' Visulization of Z-score heatmap.
 #'
-#' @import Seurat
-#' @import pheatmap
+#' @importFrom grDevices colorRampPalette
+#' @importFrom pheatmap pheatmap
 #' @export
 #'
-#' @param zscoreMatrix calculated and reshaped z-score matirx from join count analysis
+#' @param zscoreMatrix calculated and reshaped z-score matirx from join count analysis.
 #'
 #' @return Heatmap plot
+#'
 #' @examples
-#' fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
+#' fpath <- system.file("script", "humanBC.rda", package="stJoincount")
 #' load(fpath)
 #' mosaicIntegration <- rasterizeEachCluster(humanBC)
 #' joincount.result <- joincountAnalysis(mosaicIntegration)
@@ -52,9 +53,11 @@ zscoreMatrix <- function(sample, joincount.result){
 #' zscorePlot(matrix)
 
 zscorePlot <- function(zscoreMatrix){
-  pheatmap(zscoreMatrix, legend = TRUE, name = "z-score",
-                   cluster_rows=FALSE, cluster_cols=FALSE,
-                   display_numbers = TRUE, fontsize_number = 9,
-                   column_names_side = c("top"), row_names_side = c("left"), angle_col = c("0"),
-                   show_colnames = TRUE, show_rownames = TRUE, number_color = "black")
+    pheatmap(zscoreMatrix, legend = TRUE, name = "z-score",
+             cluster_rows = FALSE, cluster_cols = FALSE,
+             display_numbers = TRUE, fontsize_number = 9,
+             color=colorRampPalette(c("navy", "white", "yellow", "orange", "red"))(50),
+             column_names_side = c("top"), row_names_side = c("left"), angle_col = c("0"),
+             show_colnames = TRUE, show_rownames = TRUE, number_color = "black")
 }
+